updated CacheControl to use target-specific TemplateFactory
